move check login into fileCopy

+++ FileCopy.cs
@@ -12,8 +12,26 @@ namespace HS_FileCopy
             this._inputFilePath = inputFilePath;
             this._outputFilePath = outputFilePath;
         }
-        public bool Copy()
+        public bool StartFileCopy()
         {
+            var fileHelper = new FileHelper();
+            (bool inputExists, long inputSize) = fileHelper.FileExists(this._inputFilePath);
+            (bool outputExists, long outputSize) = fileHelper.FileExists(this._outputFilePath);
+            Console.WriteLine($"Input File - Exists: {inputExists}, Size: {inputSize} bytes\n");
+            Console.WriteLine($"Output File - Exists: {outputExists}, Size: {outputSize} bytes\n");
+
+            if (!inputExists)
+            {
+                Console.WriteLine("Input file does not exist! Aborting copy.");
+                return false;
+            }
+
+            if (outputExists)
+            {
+                Console.WriteLine("Output file already exists! Output file will be overwritten.");
+                fileHelper.DeleteFile(this._outputFilePath);
+            }
+
             this._displayCurrentTime();
             File.Copy(this._inputFilePath, this._outputFilePath);
             return true;
